requirements.txt
tensorflow==2.6.0
tensorflow-estimator==2.6.0
keras==2.6.0
matplotlib==3.4.3
pandas==1.3.2
scikit-learn==0.24.2
scikit-learn==0.24.2
pydot==1.4.2
graphviz==0.17
ann_visualizer==2.5sklearn==0.0
install==1.3.4
h5py==3.1.0
tensorboard-plugin-profile==2.5.0
opencv-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.2
scikit-image==0.18.3
pytesseract==0.3.8
pytesseract==0.3.8
loguru==0.5.3
